requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.0.3
tabulate==0.9.0
colorama==0.4.6 